csskit_proc_macro: tidy up def.rs & break out into generate.rs
